pyproject.toml
[project]
name = "agolliotapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-compressor>=4.5.1",
    "django-ninja>=1.3.0",
    "mysqlclient>=2.2.7",
    "pyfluids>=2.8.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "cursor-rules>=0.5.1",
    "djlint>=1.36.4",
]

[tool.mypy]
ignore_missing_imports = true

[tool.djlint]
profile = "django"


[tool.pyfluids]
units_system = "SIWithCelsiusAndPercents"